Add RecorderLog class to record frame grab times.

+++ test_recorder.py
@@ -29,7 +29,7 @@ if __name__ == '__main__':
 
     cam_cap = CameraCapture()
     info = cam_cap.get_framerate_info()
-    target_fps = min(24., 0.95 * info.mean_framerate)
+    target_fps = min(24., int(0.95 * info.mean_framerate))
 
     r = Recorder(args.out_file, cam_cap, fps=target_fps, auto_init=True)
     r.record()
